Module argparse::argparser

This module defines and contains all the important argument parsing functionality. The requisite types and functions are re-exported at the top-level of the crate.

Structs

ArgParseResults
This type represents the result ofparsing arguments.

ArgParser
This type represents the state and methods for parsing arguments. A new parser must be created for every set of arguments you want to parse.

Enums

ArgType
This enum represents the different types of arguments supported

Traits

ArgGetter
Represents something capable of turning a &str in the value type of your choice. Implement this to use with ArgParseResults::get_with

Functions

hashmap_parser
Function that parses Dict arguments into HashMaps. Provided for user convenience and use as an implementor of ArgGetter.

Panics

Panics if improper or no separator is found (expects key:value key2:value2...)

vec_parser
Function that parses List arguments into Vecs. Provided for user convenience and use as an implementor of ArgGetter.

Type Definitions

ParseResult
Simple type alias to reduce typing. The return type of ArgParser::parse.